visualtool-backend/requirements.txt
flask==3.1.0
flask-cors==3.0.10
pyjwt==2.8.0
alembic==1.14.1
flask-migrate==4.1.0
flask-mysqldb==2.0.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
mako==1.3.9
mysqlclient==2.2.7
sqlalchemy==2.0.38
openai==1.60.1
httpx==0.27.0
paramiko==3.5.1
PySocks==1.7.1